Uploaded a simple Python3 based driver
Python3 driver is based on the most recent protocol documentation published by SlamTec (as of Dec, 2018). This does NOT include the new ~16000 pulse rate supported by the A3 sensor, because this feature is not yet documented by SlamTec and I don't feel like trying to reverse engineer the feature out of the C++ code.